% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cancel_a_run}
\alias{cancel_a_run}
\title{Cancel A Run - Beta}
\usage{
cancel_a_run(run_id, thread_id, return_response = F)
}
\arguments{
\item{run_id}{(string) The ID of the run to cancel. Required}

\item{thread_id}{(string) The ID of the thread to which this run belongs. Required}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Cancels a run that is in_progress.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/runs/cancelRun}{Open AI Documentation}
}
