% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cancel_fine_tunedeprecated}
\alias{cancel_fine_tunedeprecated}
\title{Cancel Fine-Tunedeprecated}
\usage{
cancel_fine_tunedeprecated(fine_tune_id, return_response = F)
}
\arguments{
\item{fine_tune_id}{(string) The ID of the fine-tune job to cancel Required}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Immediately cancel a fine-tune job.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/fine-tunes/cancel}{Open AI Documentation}
}
