% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{chat_message}
\alias{chat_message}
\title{Generate a message object}
\usage{
chat_message(content, role = "user", name = NULL)
}
\arguments{
\item{content}{content of the message}

\item{role}{role of the message}

\item{name}{name of the message sender}
}
\value{
A chat_message object, which is just a named list with content, role, and name.
}
\description{
Generates a chat_message object. This is used in the chat_completion functions.
}
