% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_assistant}
\alias{create_assistant}
\title{Create Assistant - Beta}
\usage{
create_assistant(
  model,
  description = NULL,
  file_ids = NULL,
  instructions = NULL,
  metadata = NULL,
  name = NULL,
  tools = NULL,
  return_response = F
)
}
\arguments{
\item{model}{() ID of the model to use. You can use the List models API to see all of your available models, or see our Model overview for descriptions of them. Required}

\item{description}{(string or null) The description of the assistant. The maximum length is 512 characters.}

\item{file_ids}{(array) A list of file IDs attached to this assistant. There can be a maximum of 20 files attached to the assistant. Files are ordered by their creation date in ascending order.}

\item{instructions}{(string or null) The system instructions that the assistant uses. The maximum length is 32768 characters.}

\item{metadata}{(map) Set of 16 key-value pairs that can be attached to an object. This can be useful for storing additional information about the object in a structured format. Keys can be a maximum of 64 characters long and values can be a maxium of 512 characters long.}

\item{name}{(string or null) The name of the assistant. The maximum length is 256 characters.}

\item{tools}{(array) A list of tool enabled on the assistant. There can be a maximum of 128 tools per assistant. Tools can be of types code_interpreter, retrieval, or function.}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Create an assistant with a model and instructions.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/assistants/createAssistant}{Open AI Documentation}
}
