% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_assistant_file}
\alias{create_assistant_file}
\title{Create Assistant File - Beta}
\usage{
create_assistant_file(assistant_id, file_id, return_response = F)
}
\arguments{
\item{assistant_id}{(string) The ID of the assistant for which to create a File. Required}

\item{file_id}{(string) A File ID (with purpose="assistants") that the assistant should use. Useful for tools like retrieval and code_interpreter that can access files. Required}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Create an assistant file by attaching a File to an assistant.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/assistants/createAssistantFile}{Open AI Documentation}
}
