% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_embedding.R
\name{create_embedding}
\alias{create_embedding}
\title{Create embeddings}
\usage{
create_embedding(
  engine_id,
  input,
  user = NULL,
  openai_api_key = Sys.getenv("OPENAI_API_KEY"),
  openai_organization = NULL
)
}
\arguments{
\item{engine_id}{required; a length one character vector. The ID of the
engine to use for this request.}

\item{input}{required; an arbitrary length character vector. Input text to
get embeddings for, encoded as a string or array of tokens. To get embeddings
for multiple inputs in a single request, pass an array of strings or array of
token arrays. Each input must not exceed 2048 tokens in length. Unless your
are embedding code, we suggest replacing newlines (\code{\\n}) in your input
with a single space, as we have observed inferior results when newlines are
present.}

\item{user}{optional; defaults to \code{NULL}; a length one character vector.
A unique identifier representing your end-user, which will help OpenAI to
monitor and detect abuse.}

\item{openai_api_key}{required; defaults to
\code{Sys.getenv("OPENAI_API_KEY")} (i.e., the value is retrieved from the
\code{.Renviron} file); a length one character vector. Specifies OpenAI API
key.}

\item{openai_organization}{optional; defaults to \code{NULL}; a length one
character vector. Specifies OpenAI organization.}
}
\value{
Returns a list, an element of which contains embedding vector(s) for
a given input.
}
\description{
Creates an embedding vector representing the input text. See
\href{https://beta.openai.com/docs/api-reference/embeddings/create}{this page}
for details.
}
\details{
Get a vector representation of a given input that can be easily consumed by
machine learning models and algorithms. Related guide:
\href{https://beta.openai.com/docs/guides/embeddings}{Embeddings}.
}
\examples{
\dontrun{
create_embedding(
    engine_id = "text-similarity-babbage-001",
    input = c(
        "Ah, it is so boring to write documentation",
        "But examples are really crucial"
    )
)
}
}
