% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_image}
\alias{create_image}
\title{Create Image}
\usage{
create_image(
  prompt,
  n = NULL,
  response_format = NULL,
  size = NULL,
  user = NULL,
  return_response = F
)
}
\arguments{
\item{prompt}{(string) A text description of the desired image(s). The maximum length is 1000 characters. Required}

\item{n}{(integer) The number of images to generate. Must be between 1 and 10.}

\item{response_format}{(string) The format in which the generated images are returned. Must be one of url or b64_json.}

\item{size}{(string) The size of the generated images. Must be one of 256x256, 512x512, or 1024x1024.}

\item{user}{(string) A unique identifier representing your end-user, which can help OpenAI to monitor and detect abuse. Learn more.}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Creates an image given a prompt.
}
\seealso{
\href{https://beta.openai.com/docs/api-reference/images/create}{Open AI Documentation}
}
