% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_image_variation}
\alias{create_image_variation}
\title{Create Image Variation - Beta}
\usage{
create_image_variation(
  image,
  n = NULL,
  response_format = NULL,
  size = NULL,
  user = NULL,
  return_response = F
)
}
\arguments{
\item{image}{(string) The image to use as the basis for the variation(s). Must be a valid PNG file, less than 4MB, and square. Required}

\item{n}{(integer) The number of images to generate. Must be between 1 and 10.}

\item{response_format}{(string) The format in which the generated images are returned. Must be one of url or b64_json.}

\item{size}{(string) The size of the generated images. Must be one of 256x256, 512x512, or 1024x1024.}

\item{user}{(string) A unique identifier representing your end-user, which can help OpenAI to monitor and detect abuse. Learn more.}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Creates a variation of a given image.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/images/create-variation}{Open AI Documentation}
}
