% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_message}
\alias{create_message}
\title{Create Message - Beta}
\usage{
create_message(
  content,
  role,
  thread_id,
  file_ids = NULL,
  metadata = NULL,
  return_response = F
)
}
\arguments{
\item{content}{(string) The content of the message. Required}

\item{role}{(string) The role of the entity that is creating the message. Currently only user is supported. Required}

\item{thread_id}{(string) The ID of the thread to create a message for. Required}

\item{file_ids}{(array) A list of File IDs that the message should use. There can be a maximum of 10 files attached to a message. Useful for tools like retrieval and code_interpreter that can access and use files.}

\item{metadata}{(map) Set of 16 key-value pairs that can be attached to an object. This can be useful for storing additional information about the object in a structured format. Keys can be a maximum of 64 characters long and values can be a maxium of 512 characters long.}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Create a message.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/messages/createMessage}{Open AI Documentation}
}
