% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_search.R
\name{create_search}
\alias{create_search}
\title{Create search}
\usage{
create_search(
  engine_id = c("ada", "babbage", "curie", "davinci"),
  query,
  documents = NULL,
  file = NULL,
  max_rerank = 200,
  return_metadata = FALSE,
  user = NULL,
  openai_api_key = Sys.getenv("OPENAI_API_KEY"),
  openai_organization = NULL
)
}
\arguments{
\item{engine_id}{required; defaults to \code{"ada"}; a length one character
vector, one among \code{"ada"}, \code{"babbage"}, \code{"curie"}, and
\code{"davinci"}. The ID of the engine to use for this request.}

\item{query}{required; length one character vector. Query to search against
the documents.}

\item{documents}{optional; defaults to \code{NULL}; an arbitrary length
character vector. Up to 200 documents to search over. The maximum document
length (in tokens) is 2034 minus the number of tokens in the query. You
should specify either \code{documents} or a \code{file}, but not both.}

\item{file}{optional; defaults to \code{NULL}; length one character vector.
The ID of an uploaded file that contains documents to search over. You should
specify either \code{documents} or a \code{file}, but not both.}

\item{max_rerank}{required; defaults to \code{200}; a length one numeric vector
with the integer value greater than \code{0}. The maximum number of documents
to be re-ranked and returned by search. This flag only takes effect when
\code{file} is set.}

\item{return_metadata}{required; defaults to \code{FALSE}; a length one
logical vector. A special boolean flag for showing metadata. If set to
\code{TRUE}, each document entry in the returned JSON will contain a
"metadata" field. This flag only takes effect when \code{file} is set.}

\item{user}{optional; defaults to \code{NULL}; a length one character vector.
A unique identifier representing your end-user, which will help OpenAI to
monitor and detect abuse.}

\item{openai_api_key}{required; defaults to
\code{Sys.getenv("OPENAI_API_KEY")} (i.e., the value is retrieved from the
\code{.Renviron} file); a length one character vector. Specifies OpenAI API
key.}

\item{openai_organization}{optional; defaults to \code{NULL}; a length one
character vector. Specifies OpenAI organization.}
}
\value{
Returns a list, elements of which contain score of each document and
supplementary information.
}
\description{
Given a query and a set of documents or labels, the model ranks each document
based on its semantic similarity to the provided query. See
\href{https://beta.openai.com/docs/api-reference/searches/create}{this page}
for details.
}
\examples{
\dontrun{
create_search(
    documents = c("White House", "hospital", "school"),
    query = "the president"
)
}
}
