% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_speech}
\alias{create_speech}
\title{Create Speech}
\usage{
create_speech(
  input,
  model,
  voice,
  response_format = NULL,
  speed = NULL,
  return_response = F
)
}
\arguments{
\item{input}{(string) The text to generate audio for. The maximum length is 4096 characters. Required}

\item{model}{(string) One of the available TTS models: tts-1 or tts-1-hd Required}

\item{voice}{(string) The voice to use when generating the audio. Supported voices are alloy, echo, fable, onyx, nova, and shimmer. Required}

\item{response_format}{(string) The format to audio in. Supported formats are mp3, opus, aac, and flac.}

\item{speed}{(number) The speed of the generated audio. Select a value from 0.25 to 4.0. 1.0 is the default.}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Generates audio from the input text.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/audio/createSpeech}{Open AI Documentation}
}
