% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_thread}
\alias{create_thread}
\title{Create Thread - Beta}
\usage{
create_thread(messages = NULL, metadata = NULL, return_response = F)
}
\arguments{
\item{messages}{(array) A list of messages to start the thread with.}

\item{metadata}{(map) Set of 16 key-value pairs that can be attached to an object. This can be useful for storing additional information about the object in a structured format. Keys can be a maximum of 64 characters long and values can be a maxium of 512 characters long.}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Create a thread.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/threads/createThread}{Open AI Documentation}
}
