% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_thread_and_run}
\alias{create_thread_and_run}
\title{Create Thread And Run - Beta}
\usage{
create_thread_and_run(
  assistant_id,
  instructions = NULL,
  metadata = NULL,
  model = NULL,
  tools = NULL,
  return_response = F
)
}
\arguments{
\item{assistant_id}{(string) The ID of the assistant to use to execute this run. Required}

\item{instructions}{(string or null) Override the default system message of the assistant. This is useful for modifying the behavior on a per-run basis.}

\item{metadata}{(map) Set of 16 key-value pairs that can be attached to an object. This can be useful for storing additional information about the object in a structured format. Keys can be a maximum of 64 characters long and values can be a maxium of 512 characters long.}

\item{model}{(string or null) The ID of the Model to be used to execute this run. If a value is provided here, it will override the model associated with the assistant. If not, the model associated with the assistant will be used.}

\item{tools}{(array or null) Override the tools the assistant can use for this run. This is useful for modifying the behavior on a per-run basis.}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Create a thread and run it in one request.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/runs/createThreadAndRun}{Open AI Documentation}
}
