% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_transcription}
\alias{create_transcription}
\title{Create Transcription - Beta}
\usage{
create_transcription(
  file,
  model,
  language = NULL,
  prompt = NULL,
  response_format = NULL,
  temperature = NULL,
  return_response = F
)
}
\arguments{
\item{file}{(string) The audio file to transcribe, in one of these formats: mp3, mp4, mpeg, mpga, m4a, wav, or webm. Required}

\item{model}{(string) ID of the model to use. Only whisper-1 is currently available. Required}

\item{language}{(string) The language of the input audio. Supplying the input language in ISO-639-1 format will improve accuracy and latency.}

\item{prompt}{(string) An optional text to guide the model's style or continue a previous audio segment. The prompt should match the audio language.}

\item{response_format}{(string) The format of the transcript output, in one of these options: json, text, srt, verbose_json, or vtt.}

\item{temperature}{(number) The sampling temperature, between 0 and 1. Higher values like 0.8 will make the output more random, while lower values like 0.2 will make it more focused and deterministic. If set to 0, the model will use log probability to automatically increase the temperature until certain thresholds are hit.}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Transcribes audio into the input language.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/audio/create}{Open AI Documentation}
}
