% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_translation}
\alias{create_translation}
\title{Create Translation}
\usage{
create_translation(
  file,
  model,
  prompt = NULL,
  response_format = NULL,
  temperature = NULL,
  return_response = F
)
}
\arguments{
\item{file}{(file) The audio file object (not file name) translate, in one of these formats: flac, mp3, mp4, mpeg, mpga, m4a, ogg, wav, or webm. Required}

\item{model}{(string) ID of the model to use. Only whisper-1 is currently available. Required}

\item{prompt}{(string) An optional text to guide the model's style or continue a previous audio segment. The prompt should be in English.}

\item{response_format}{(string) The format of the transcript output, in one of these options: json, text, srt, verbose_json, or vtt.}

\item{temperature}{(number) The sampling temperature, between 0 and 1. Higher values like 0.8 will make the output more random, while lower values like 0.2 will make it more focused and deterministic. If set to 0, the model will use log probability to automatically increase the temperature until certain thresholds are hit.}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Translates audio into English.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/audio/createTranslation}{Open AI Documentation}
}
