% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{delete_assistant}
\alias{delete_assistant}
\title{Delete Assistant - Beta}
\usage{
delete_assistant(assistant_id, return_response = F)
}
\arguments{
\item{assistant_id}{(string) The ID of the assistant to delete. Required}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Delete an assistant.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/assistants/deleteAssistant}{Open AI Documentation}
}
