% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{delete_assistant_file}
\alias{delete_assistant_file}
\title{Delete Assistant File - Beta}
\usage{
delete_assistant_file(assistant_id, file_id, return_response = F)
}
\arguments{
\item{assistant_id}{(string) The ID of the assistant that the file belongs to. Required}

\item{file_id}{(string) The ID of the file to delete. Required}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Delete an assistant file.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/assistants/deleteAssistantFile}{Open AI Documentation}
}
