% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{delete_file}
\alias{delete_file}
\title{Delete File}
\usage{
delete_file(file_id, return_response = F)
}
\arguments{
\item{file_id}{(string) The ID of the file to use for this request. Required}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Delete a file.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/files/delete}{Open AI Documentation}
}
