% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{delete_thread}
\alias{delete_thread}
\title{Delete Thread - Beta}
\usage{
delete_thread(thread_id, return_response = F)
}
\arguments{
\item{thread_id}{(string) The ID of the thread to delete. Required}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Delete a thread.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/threads/deleteThread}{Open AI Documentation}
}
