% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{get_embedding}
\alias{get_embedding}
\title{Get Embeddings}
\usage{
get_embedding(
  input,
  model = "text-embedding-3-small",
  dimensions = NULL,
  user = NULL,
  ...,
  credentials = default_credentials(wrapper),
  action = default_action,
  decode_if_success = decode_if_success_default_value
)
}
\arguments{
\item{input}{The input text}

\item{model}{Which embedding model to use}

\item{dimensions}{How many dimensions to return}

\item{user}{A user ID}

\item{...}{Arguments passed to the request function}

\item{credentials}{Credentials}

\item{action}{perform or dry run}

\item{decode_if_success}{decode if success?}
}
\value{
an embedding object
}
\description{
Gets embeddings from openai
}
