% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai.R
\docType{data}
\name{iso_languages}
\alias{iso_languages}
\title{ISO 639-1 two-letter language codes}
\description{
A named vector of two-letter language codes.
}
\references{
\url{https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes}
}
\keyword{data}
