% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_engines.R
\name{list_engines}
\alias{list_engines}
\title{List Engines}
\usage{
list_engines(
  openai_api_key = Sys.getenv("OPENAI_API_KEY"),
  openai_organization = NULL
)
}
\arguments{
\item{openai_api_key}{a length one character vector containing OpenAI API
key. Defaults to \code{Sys.getenv("OPENAI_API_KEY")} (i.e., the value is
retrieved from the \code{.Renviron} file).}

\item{openai_organization}{an optional length one character vector specifying
OpenAI organization. Defaults to \code{NULL}.}
}
\value{
A list of engines.
}
\description{
Lists the currently available engines, and provides basic information about
each one such as the owner and availability.
}
\examples{
\dontrun{
list_engines()
}

}
