% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_files}
\alias{list_files}
\title{List Files
Returns a list of files that belong to the user's organization.}
\usage{
list_files(return_response = F)
}
\description{
List Files
Returns a list of files that belong to the user's organization.
}
