% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_fine_tunesdeprecated}
\alias{list_fine_tunesdeprecated}
\title{List Fine-Tunesdeprecated}
\usage{
list_fine_tunesdeprecated(return_response = F)
}
\arguments{
\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
List your organization's fine-tuning jobs
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/fine-tunes/list}{Open AI Documentation}
}
