% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_fine_tuning_events}
\alias{list_fine_tuning_events}
\title{List Fine-Tuning Events}
\usage{
list_fine_tuning_events(
  fine_tuning_job_id,
  after = NULL,
  limit = NULL,
  return_response = F
)
}
\arguments{
\item{fine_tuning_job_id}{(string) The ID of the fine-tuning job to get events for. Required}

\item{after}{(string) Identifier for the last event from the previous pagination request.}

\item{limit}{(integer) Number of events to retrieve.}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Get status updates for a fine-tuning job.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/fine-tuning/list-events}{Open AI Documentation}
}
