% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_fine_tuning_jobs}
\alias{list_fine_tuning_jobs}
\title{List Fine-Tuning Jobs}
\usage{
list_fine_tuning_jobs(after = NULL, limit = NULL, return_response = F)
}
\arguments{
\item{after}{(string) Identifier for the last job from the previous pagination request.}

\item{limit}{(integer) Number of fine-tuning jobs to retrieve.}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
List your organization's fine-tuning jobs
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/fine-tuning/list}{Open AI Documentation}
}
