% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_messages}
\alias{list_messages}
\title{List Messages - Beta}
\usage{
list_messages(
  thread_id,
  after = NULL,
  before = NULL,
  limit = NULL,
  order = NULL,
  return_response = F
)
}
\arguments{
\item{thread_id}{(string) The ID of the thread the messages belong to. Required}

\item{after}{(string) A cursor for use in pagination. after is an object ID that defines your place in the list. For instance, if you make a list request and receive 100 objects, ending with obj_foo, your subsequent call can include after=obj_foo in order to fetch the next page of the list.}

\item{before}{(string) A cursor for use in pagination. before is an object ID that defines your place in the list. For instance, if you make a list request and receive 100 objects, ending with obj_foo, your subsequent call can include before=obj_foo in order to fetch the previous page of the list.}

\item{limit}{(integer) A limit on the number of objects to be returned. Limit can range between 1 and 100, and the default is 20.}

\item{order}{(string) Sort order by the created_at timestamp of the objects. asc for ascending order and desc for descending order.}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Returns a list of messages for a given thread.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/messages/listMessages}{Open AI Documentation}
}
