% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{list_models}
\alias{list_models}
\title{List models}
\usage{
list_models(
  .credentials = get_credentials_from_wrapper(wrapper),
  .perform = perform_by_default,
  .extract = extract_body_by_default,
  .extractor = extractor
)
}
\arguments{
\item{...}{arguments passed to the perform function}

\item{credentials}{API token}

\item{action}{either "perform" or "dryrun"}

\item{decode_if_success}{should decode if success?}

\item{post_process}{if true, convert return object to tibble}
}
\value{
A list of available models
}
\description{
Lists available models from OpenAI
}
