% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{modify_message}
\alias{modify_message}
\title{Modify Message - Beta}
\usage{
modify_message(message_id, thread_id, metadata = NULL, return_response = F)
}
\arguments{
\item{message_id}{(string) The ID of the message to modify. Required}

\item{thread_id}{(string) The ID of the thread to which this message belongs. Required}

\item{metadata}{(map) Set of 16 key-value pairs that can be attached to an object. This can be useful for storing additional information about the object in a structured format. Keys can be a maximum of 64 characters long and values can be a maxium of 512 characters long.}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Modifies a message.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/messages/modifyMessage}{Open AI Documentation}
}
