% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{quick_chat_completion}
\alias{quick_chat_completion}
\title{Quick chat completion}
\usage{
quick_chat_completion(user_message, system_message = NULL, ...)
}
\arguments{
\item{user_message}{User message}

\item{system_message}{System message}

\item{...}{arguments passed to chat_completion}
}
\value{
The returned completion from OpenAI
}
\description{
Generates a quick chat completion
}
