% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{retrieve_assistant}
\alias{retrieve_assistant}
\title{Retrieve Assistant - Beta}
\usage{
retrieve_assistant(assistant_id, return_response = F)
}
\arguments{
\item{assistant_id}{(string) The ID of the assistant to retrieve. Required}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Retrieves an assistant.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/assistants/getAssistant}{Open AI Documentation}
}
