% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{retrieve_assistant_file}
\alias{retrieve_assistant_file}
\title{Retrieve Assistant File - Beta}
\usage{
retrieve_assistant_file(assistant_id, file_id, return_response = F)
}
\arguments{
\item{assistant_id}{(string) The ID of the assistant who the file belongs to. Required}

\item{file_id}{(string) The ID of the file we're getting. Required}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Retrieves an AssistantFile.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/assistants/getAssistantFile}{Open AI Documentation}
}
