% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_engine.R
\name{retrieve_engine}
\alias{retrieve_engine}
\title{Retrieve engine}
\usage{
retrieve_engine(
  engine_id,
  openai_api_key = Sys.getenv("OPENAI_API_KEY"),
  openai_organization = NULL
)
}
\arguments{
\item{engine_id}{required; a length one character vector. The ID of the
engine to use for this request.}

\item{openai_api_key}{required; defaults to
\code{Sys.getenv("OPENAI_API_KEY")} (i.e., the value is retrieved from the
\code{.Renviron} file); a length one character vector. Specifies OpenAI API
key.}

\item{openai_organization}{optional; defaults to \code{NULL}; a length one
character vector. Specifies OpenAI organization.}
}
\value{
Returns a list, elements of which contain information about the
engine.
}
\description{
Retrieves an engine instance, providing basic information about the engine
such as the owner and availability. See
\href{https://beta.openai.com/docs/api-reference/engines/retrieve}{this page}
for details.
}
\details{
Engines describe and provide access to the various models available in the
API. You can refer to the \href{https://beta.openai.com/docs/engines}{Engines}
documentation to understand what engines are available and the differences
between them.
}
\examples{
\dontrun{
retrieve_engine("text-davinci-002")
}
}
\seealso{
Other engine functions: 
\code{\link{list_engines}()}
}
\concept{engine functions}
