% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{retrieve_file_content}
\alias{retrieve_file_content}
\title{Retrieve File Content
Returns the contents of the specified file}
\usage{
retrieve_file_content(file_id, return_response = F)
}
\arguments{
\item{file_id}{(string) The ID of the file to use for this request Required}
}
\description{
Retrieve File Content
Returns the contents of the specified file
}
