% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{retrieve_fine_tune}
\alias{retrieve_fine_tune}
\title{Retrieve Fine-Tune - Beta
Gets info about the fine-tune job.}
\usage{
retrieve_fine_tune(fine_tune_id, return_response = F)
}
\arguments{
\item{fine_tune_id}{(string) The ID of the fine-tune job Required}
}
\description{
Retrieve Fine-Tune - Beta
Gets info about the fine-tune job.
}
