% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{retrieve_fine_tuning_job}
\alias{retrieve_fine_tuning_job}
\title{Retrieve Fine-Tuning Job}
\usage{
retrieve_fine_tuning_job(fine_tuning_job_id, return_response = F)
}
\arguments{
\item{fine_tuning_job_id}{(string) The ID of the fine-tuning job. Required}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Get info about a fine-tuning job.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/fine-tuning/retrieve}{Open AI Documentation}
}
