% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{retrieve_message}
\alias{retrieve_message}
\title{Retrieve Message - Beta}
\usage{
retrieve_message(message_id, thread_id, return_response = F)
}
\arguments{
\item{message_id}{(string) The ID of the message to retrieve. Required}

\item{thread_id}{(string) The ID of the thread to which this message belongs. Required}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Retrieve a message.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/messages/getMessage}{Open AI Documentation}
}
