% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{retrieve_message_file}
\alias{retrieve_message_file}
\title{Retrieve Message File - Beta}
\usage{
retrieve_message_file(file_id, message_id, thread_id, return_response = F)
}
\arguments{
\item{file_id}{(string) The ID of the file being retrieved. Required}

\item{message_id}{(string) The ID of the message the file belongs to. Required}

\item{thread_id}{(string) The ID of the thread to which the message and File belong. Required}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Retrieves a message file.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/messages/getMessageFile}{Open AI Documentation}
}
