% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{retrieve_run}
\alias{retrieve_run}
\title{Retrieve Run - Beta}
\usage{
retrieve_run(run_id, thread_id, return_response = F)
}
\arguments{
\item{run_id}{(string) The ID of the run to retrieve. Required}

\item{thread_id}{(string) The ID of the thread that was run. Required}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Retrieves a run.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/runs/getRun}{Open AI Documentation}
}
