% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{retrieve_run_step}
\alias{retrieve_run_step}
\title{Retrieve Run Step - Beta}
\usage{
retrieve_run_step(run_id, step_id, thread_id, return_response = F)
}
\arguments{
\item{run_id}{(string) The ID of the run to which the run step belongs. Required}

\item{step_id}{(string) The ID of the run step to retrieve. Required}

\item{thread_id}{(string) The ID of the thread to which the run and run step belongs. Required}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Retrieves a run step.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/runs/getRunStep}{Open AI Documentation}
}
