% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{retrieve_thread}
\alias{retrieve_thread}
\title{Retrieve Thread - Beta}
\usage{
retrieve_thread(thread_id, return_response = F)
}
\arguments{
\item{thread_id}{(string) The ID of the thread to retrieve. Required}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Retrieves a thread.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/threads/getThread}{Open AI Documentation}
}
