% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{set_api_key}
\alias{set_api_key}
\title{Save an API key to an environment variable}
\usage{
set_api_key(credentials = ask_for_credentials(wrapper))
}
\arguments{
\item{credentials}{the value of the key to save. Asks the user by default.}
}
\value{
A logical vector as per Sys.setenv
}
\description{
Saves the key
}
