% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{submit_tool_outputs_to_run}
\alias{submit_tool_outputs_to_run}
\title{Submit Tool Outputs To Run - Beta}
\usage{
submit_tool_outputs_to_run(
  run_id,
  thread_id,
  tool_outputs,
  return_response = F
)
}
\arguments{
\item{run_id}{(string) The ID of the run that requires the tool output submission. Required}

\item{thread_id}{(string) The ID of the thread to which this run belongs. Required}

\item{tool_outputs}{(array) A list of tools for which the outputs are being submitted. Required}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
When a run has the status: "requires_action" and required_action.type is submit_tool_outputs, this endpoint can be used to submit the outputs from the tool calls once they're all completed. All outputs must be submitted in a single request.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/runs/submitToolOutputs}{Open AI Documentation}
}
