% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_file.R
\name{upload_file}
\alias{upload_file}
\title{Upload file}
\usage{
upload_file(
  file,
  purpose = c("search", "answers", "classifications", "fine-tune"),
  openai_api_key = Sys.getenv("OPENAI_API_KEY"),
  openai_organization = NULL
)
}
\arguments{
\item{file}{required; a length one character vector. Name of the JSON Lines
file to be uploaded. If the \code{purpose} is set to "search" or "answers", each
line is a JSON record with a "text" field and an optional "metadata" field.
Only "text" field will be used for search. Specially, when the \code{purpose} is
"answers", "\\n" is used as a delimiter to chunk contents in the "text"
field into multiple documents for finer-grained matching. If the \code{purpose}
is set to "classifications", each line is a JSON record representing a
single training example with "text" and "label" fields along with an
optional "metadata" field. If the \code{purpose} is set to "fine-tune", each
line is a JSON record with "prompt" and "completion" fields representing
your \href{https://beta.openai.com/docs/guides/fine-tuning/prepare-training-data}{training examples}.}

\item{purpose}{required; defaults to \code{"search"}; a length one character
vector, one among \code{"search"}, \code{"answers"}, \code{"classifications"}, and
\code{"fine-tune"}. The intended purpose of the uploaded documents. This allows
us to validate the format of the uploaded file.}

\item{openai_api_key}{required; defaults to \code{Sys.getenv("OPENAI_API_KEY")}
(i.e., the value is retrieved from the \code{.Renviron} file); a length one
character vector. Specifies OpenAI API key.}

\item{openai_organization}{optional; defaults to \code{NULL}; a length one
character vector. Specifies OpenAI organization.}
}
\value{
Returns a list, elements of which contains ID of the uploaded file
and other supplementary information.
}
\description{
Upload a file that contains document(s) to be used across various
endpoints/features. Currently, the size of all the files uploaded by one
organization can be up to 1 GB. Please contact us if you need to increase the
storage limit. See \href{https://beta.openai.com/docs/api-reference/files/upload}{this page} for details.
}
\details{
Files are used to upload documents that can be used across features like
\href{https://beta.openai.com/docs/api-reference/answers}{Answers},
\href{https://beta.openai.com/docs/api-reference/searches}{Search}, and
\href{https://beta.openai.com/docs/api-reference/classifications}{Classifications}.
}
\examples{
\dontrun{
file <- system.file("extdata", "classification-file.jsonl", package = "openai")
upload_file(file = file, purpose = "classification")
}
}
\seealso{
Other file functions: 
\code{\link{delete_file}()},
\code{\link{list_files}()},
\code{\link{retrieve_file_content}()},
\code{\link{retrieve_file}()}
}
\concept{file functions}
