% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_answer}
\alias{create_answer}
\title{Create Answer - Beta}
\usage{
create_answer(
  examples,
  examples_context,
  model,
  question,
  documents = NULL,
  expand = NULL,
  file = NULL,
  logit_bias = NULL,
  logprobs = NULL,
  max_rerank = NULL,
  max_tokens = NULL,
  n = NULL,
  return_metadata = NULL,
  return_prompt = NULL,
  search_model = NULL,
  stop = NULL,
  temperature = NULL,
  user = NULL,
  return_response = F
)
}
\arguments{
\item{examples}{(array) List of (question, answer) pairs that will help steer the model towards the tone and answer format you'd like. We recommend adding 2 to 3 examples. Required}

\item{examples_context}{(string) A text snippet containing the contextual information used to generate the answers for the examples you provide. Required}

\item{model}{(string) ID of the engine to use for completion. You can select one of ada, babbage, curie, or davinci. Required}

\item{question}{(string) Question to get answered. Required}

\item{documents}{(array) List of documents from which the answer for the input question should be derived. If this is an empty list, the question will be answered based on the question-answer examples. You should specify either documents or a file, but not both.}

\item{expand}{(array) If an object name is in the list, we provide the full information of the object; otherwise, we only provide the object ID. Currently we support completion and file objects for expansion.}

\item{file}{(string) The ID of an uploaded file that contains documents to search over. See upload file for how to upload a file of the desired format and purpose. You should specify either documents or a file, but not both.}

\item{logit_bias}{(map) Modify the likelihood of specified tokens appearing in the completion. Accepts a json object that maps tokens (specified by their token ID in the GPT tokenizer) to an associated bias value from -100 to 100. You can use this tokenizer tool (which works for both GPT-2 and GPT-3) to convert text to token IDs. Mathematically, the bias is added to the logits generated by the model prior to sampling. The exact effect will vary per model, but values between -1 and 1 should decrease or increase likelihood of selection; values like -100 or 100 should result in a ban or exclusive selection of the relevant token. As an example, you can pass {"50256": -100} to prevent the <|endoftext|> token from being generated.}

\item{logprobs}{(integer) Include the log probabilities on the logprobs most likely tokens, as well the chosen tokens. For example, if logprobs is 5, the API will return a list of the 5 most likely tokens. The API will always return the logprob of the sampled token, so there may be up to logprobs+1 elements in the response. The maximum value for logprobs is 5. If you need more than this, please contact support@openai.com and describe your use case. When logprobs is set, completion will be automatically added into expand to get the logprobs.}

\item{max_rerank}{(integer) The maximum number of documents to be ranked by Search when using file. Setting it to a higher value leads to improved accuracy but with increased latency and cost.}

\item{max_tokens}{(integer) The maximum number of tokens allowed for the generated answer}

\item{n}{(integer) How many answers to generate for each question.}

\item{return_metadata}{(boolean) A special boolean flag for showing metadata. If set to true, each document entry in the returned JSON will contain a "metadata" field. This flag only takes effect when file is set.}

\item{return_prompt}{(boolean) If set to true, the returned JSON will include a "prompt" field containing the final prompt that was used to request a completion. This is mainly useful for debugging purposes.}

\item{search_model}{(string) ID of the engine to use for Search. You can select one of ada, babbage, curie, or davinci.}

\item{stop}{(string or array) Up to 4 sequences where the API will stop generating further tokens. The returned text will not contain the stop sequence.}

\item{temperature}{(number) What sampling temperature to use. Higher values mean the model will take more risks and value 0 (argmax sampling) works better for scenarios with a well-defined answer.}

\item{user}{(string) A unique identifier representing your end-user, which will help OpenAI to monitor and detect abuse.}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Answers the specified question using the provided documents and examples.
}
\seealso{
\href{https://beta.openai.com/docs/api-reference/answers/create}{Open AI Documentation}
}
