% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_edit}
\alias{create_edit}
\title{Create Edit}
\usage{
create_edit(
  instruction,
  model,
  input = NULL,
  n = NULL,
  temperature = NULL,
  top_p = NULL,
  return_response = F
)
}
\arguments{
\item{instruction}{(string) The instruction that tells the model how to edit the prompt. Required}

\item{model}{(string) ID of the model to use. You can use the List models API to see all of your available models, or see our Model overview for descriptions of them. Required}

\item{input}{(string) The input text to use as a starting point for the edit.}

\item{n}{(integer) How many edits to generate for the input and instruction.}

\item{temperature}{(number) What sampling temperature to use. Higher values means the model will take more risks. Try 0.9 for more creative applications, and 0 (argmax sampling) for ones with a well-defined answer. We generally recommend altering this or top_p but not both.}

\item{top_p}{(number) An alternative to sampling with temperature, called nucleus sampling, where the model considers the results of the tokens with top_p probability mass. So 0.1 means only the tokens comprising the top 10\% probability mass are considered. We generally recommend altering this or temperature but not both.}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Creates a new edit for the provided input, instruction, and parameters
}
\seealso{
\href{https://beta.openai.com/docs/api-reference/edits/create}{Open AI Documentation}
}
