% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_embedding}
\alias{create_embedding}
\title{Create Embedding - Beta}
\usage{
create_embedding(engine_id, input, return_response = F)
}
\arguments{
\item{input}{(string or array) Input text to get embeddings for, encoded as a string or array of tokens. To get embeddings for multiple inputs in a single request, pass and array of strings or array of token arrays. Each input must not exceed 2048 tokens in length.
We suggest replacing newlines ("\n") in your input with a single space, as we have observed inferior results when newlines are present.}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}

\item{enging_id}{(string) The ID of the engine to use for this request Required}
}
\description{
Creates an embedding vector representing the input text.
}
\seealso{
\href{https://beta.openai.com/docs/api-reference/fine-tunes/cancel}{Open AI Documentation}
}
