% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_search}
\alias{create_search}
\title{Create Search}
\usage{
create_search(
  engine_id,
  query,
  documents = NULL,
  file = NULL,
  max_rerank = NULL,
  return_metadata = NULL,
  return_response = F
)
}
\arguments{
\item{engine_id}{(string) The ID of the engine to use for this request.  You can select one of ada, babbage, curie, or davinci. Required}

\item{query}{(string) Query to search against the documents. Required}

\item{documents}{(array) Up to 200 documents to search over, provided as a list of strings. The maximum document length (in tokens) is 2034 minus the number of tokens in the query. You should specify either documents or a file, but not both.}

\item{file}{(string) The ID of an uploaded file that contains documents to search over. You should specify either documents or a file, but not both.}

\item{max_rerank}{(integer) The maximum number of documents to be re-ranked and returned by search. This flag only takes effect when file is set.}

\item{return_metadata}{(boolean) A special boolean flag for showing metadata. If set to true, each document entry in the returned JSON will contain a "metadata" field. This flag only takes effect when file is set.}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
The search endpoint computes similarity scores between provided query and documents. Documents can be passed directly to the API if there are no more than 200 of them.
}
\seealso{
\href{https://beta.openai.com/docs/api-reference/searches/create}{Open AI Documentation}
}
