% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{delete_file}
\alias{delete_file}
\title{Delete File
Delete a file.}
\usage{
delete_file(file_id, return_response = F)
}
\arguments{
\item{file_id}{(string) The ID of the file to use for this request Required}
}
\description{
Delete File
Delete a file.
}
