% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{delete_fine_tune_model}
\alias{delete_fine_tune_model}
\title{Delete Fine-Tune Model}
\usage{
delete_fine_tune_model(model, return_response = F)
}
\arguments{
\item{model}{(string) The model to delete Required}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Delete a fine-tuned model. You must have the Owner role in your organization.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/fine-tunes/delete-model}{Open AI Documentation}
}
