% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_engines.R
\name{list_engines}
\alias{list_engines}
\title{List engines}
\usage{
list_engines(
  openai_api_key = Sys.getenv("OPENAI_API_KEY"),
  openai_organization = NULL
)
}
\arguments{
\item{openai_api_key}{required; defaults to \code{Sys.getenv("OPENAI_API_KEY")}
(i.e., the value is retrieved from the \code{.Renviron} file); a length one
character vector. Specifies OpenAI API key.}

\item{openai_organization}{optional; defaults to \code{NULL}; a length one
character vector. Specifies OpenAI organization.}
}
\value{
Returns a list, an element of which is a data frame containing
information about engines.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\strong{Note:} This endpoint is deprecated and soon will be removed. Please use
its replacement,
\href{https://platform.openai.com/docs/api-reference/models}{Models}, instead. The
replacement function in this package is \code{list_models()}.

Lists available engines and provides basic information about such engines.
See \href{https://platform.openai.com/docs/api-reference/engines/list}{this page}
for details.
}
\details{
For arguments description please refer to the \href{https://platform.openai.com/docs/api-reference/engines/list}{official documentation}.
}
\examples{
\dontrun{
list_engines()
# ->
list_models()
}
}
\seealso{
Other engine functions: 
\code{\link{retrieve_engine}()}
}
\concept{engine functions}
\keyword{internal}
