% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_engines}
\alias{list_engines}
\title{List Engines}
\usage{
list_engines(return_response = F)
}
\arguments{
\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Lists the currently available engines, and provides basic information about each one such as the owner and availability.
}
\seealso{
\href{https://beta.openai.com/docs/api-reference/engines/list}{Open AI Documentation}
}
