% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_engines}
\alias{list_engines}
\title{List Engines
Lists the currently available engines, and provides basic information about each one such as the owner and availability.}
\usage{
list_engines(return_response = F)
}
\description{
List Engines
Lists the currently available engines, and provides basic information about each one such as the owner and availability.
}
