% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_files}
\alias{list_files}
\title{List Files}
\usage{
list_files(purpose = NULL, return_response = F)
}
\arguments{
\item{purpose}{(string) Only return files with the given purpose.}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Returns a list of files that belong to the user's organization.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/files/list}{Open AI Documentation}
}
