% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_fine_tunes}
\alias{list_fine_tunes}
\title{List Fine-Tunes - Beta
List your organization's fine-tuning jobs}
\usage{
list_fine_tunes(return_response = F)
}
\description{
List Fine-Tunes - Beta
List your organization's fine-tuning jobs
}
