% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{list_models}
\alias{list_models}
\title{List models}
\usage{
list_models(
  ...,
  credentials = default_credentials(wrapper),
  action = default_action,
  decode_if_success = decode_if_success_default_value
)
}
\arguments{
\item{...}{arguments passed to the perform function}

\item{credentials}{API token}

\item{action}{either "perform" or "dryrun"}

\item{decode_if_success}{should decode if success?}
}
\value{
A list of available models
}
\description{
Lists available models from OpenAI
}
