% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_models}
\alias{list_models}
\title{List Models}
\usage{
list_models(return_response = F)
}
\arguments{
\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Lists the currently available models, and provides basic information about each one such as the owner and availability.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/models/list}{Open AI Documentation}
}
