% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{retrieve_engine}
\alias{retrieve_engine}
\title{Retrieve Engine
Retrieves an engine instance, providing basic information about the engine such as the owner and availability.}
\usage{
retrieve_engine(engine_id, return_response = F)
}
\arguments{
\item{engine_id}{(string) The ID of the engine to use for this request Required}
}
\description{
Retrieve Engine
Retrieves an engine instance, providing basic information about the engine such as the owner and availability.
}
