% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{retrieve_file}
\alias{retrieve_file}
\title{Retrieve File}
\usage{
retrieve_file(file_id, return_response = F)
}
\arguments{
\item{file_id}{(string) The ID of the file to use for this request Required}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Returns information about a specific file.
}
\seealso{
\href{https://beta.openai.com/docs/api-reference/files/retrieve}{Open AI Documentation}
}
