% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{retrieve_file}
\alias{retrieve_file}
\title{Retrieve File
Returns information about a specific file.}
\usage{
retrieve_file(file_id, return_response = F)
}
\arguments{
\item{file_id}{(string) The ID of the file to use for this request Required}
}
\description{
Retrieve File
Returns information about a specific file.
}
