% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{retrieve_fine_tune}
\alias{retrieve_fine_tune}
\title{Retrieve Fine-Tune - Beta}
\usage{
retrieve_fine_tune(fine_tune_id, return_response = F)
}
\arguments{
\item{fine_tune_id}{(string) The ID of the fine-tune job Required}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Gets info about the fine-tune job.
}
\seealso{
\href{https://beta.openai.com/docs/api-reference/fine-tunes/retrieve}{Open AI Documentation}
}
