% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{retrieve_model}
\alias{retrieve_model}
\title{Retrieve model}
\usage{
retrieve_model(
  model,
  ...,
  credentials = default_credentials(wrapper),
  action = default_action,
  decode_if_success = decode_if_success_default_value
)
}
\arguments{
\item{model}{Model ID}

\item{...}{arguments passed to the request performer}

\item{credentials}{API key}

\item{action}{perform or dry run}

\item{decode_if_success}{decode the result?}
}
\value{
List of information about a model instance.
}
\description{
Retrieve a model from the id
}
