% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{upload_file}
\alias{upload_file}
\title{Upload File}
\usage{
upload_file(file, purpose, return_response = F)
}
\arguments{
\item{file}{(string) The File object (not file name) to be uploaded. Required}

\item{purpose}{(string) The intended purpose of the uploaded file. Use "fine-tune" for Fine-tuning and "assistants" for Assistants and Messages. This allows us to validate the format of the uploaded file is correct for fine-tuning. Required}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Upload a file that can be used across various endpoints/features. The size of all the files uploaded by one organization can be up to 100 GB.
}
\seealso{
\href{https://platform.openai.com/docs/api-reference/files/create}{Open AI Documentation}
}
