% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{upload_file}
\alias{upload_file}
\title{Upload File}
\usage{
upload_file(file, purpose, return_response = F)
}
\arguments{
\item{file}{(string) Name of the JSON Lines file to be uploaded. If the purpose is set to "fine-tune", each line is a JSON record with "prompt" and "completion" fields representing your training examples. Required}

\item{purpose}{(string) The intended purpose of the uploaded documents. Use "fine-tune" for Fine-tuning. This allows us to validate the format of the uploaded file. Required}

\item{return_response}{(boolean) Whether to return the API response or parse the contents of the response. Defaults to FALSE (parse the response).}
}
\description{
Upload a file that contains document(s) to be used across various endpoints/features. Currently, the size of all the files uploaded by one organization can be up to 1 GB. Please contact us if you need to increase the storage limit.
}
\seealso{
\href{https://beta.openai.com/docs/api-reference/files/upload}{Open AI Documentation}
}
